django-react-sandbag/backend/requirements.txt

django==3.1.3
djangorestframework==3.11.2 # version compatible with Djoser
markdown      
django-filter  
djoser
social-auth-app-django
